[feat] physics-2d supports switching back-end in EDITOR
