+++ cocos/physics-2d/framework/components/colliders/collider-2d.ts
@@ -7,7 +7,7 @@ import { PhysicsGroup } from '../../../../physics/framework/physics-enum';
 import { Eventify } from '../../../../core/event';
 
 import { RigidBody2D } from '../rigid-body-2d';
-import { createShape } from '../../instance';
+import { createShape } from '../../physics-selector';
 import { ECollider2DType } from '../../physics-types';
 import { IBaseShape } from '../../../spec/i-physics-shape';
 import { legacyCC } from '../../../../core/global-exports';
