+++ cocos/physics-2d/framework/components/rigid-body-2d.ts
@@ -3,7 +3,7 @@ import { IRigidBody2D } from '../../spec/i-rigid-body';
 import { _decorator, Vec2, IVec2Like } from '../../../core';
 import { ERigidBody2DType } from '../physics-types';
 import { ccclass } from '../../../core/data/class-decorator';
-import { createRigidBody } from '../instance';
+import { createRigidBody } from '../physics-selector';
 import { PhysicsGroup } from '../../../physics/framework/physics-enum';
 import { legacyCC } from '../../../core/global-exports';
 import { Component } from '../../../scene-graph';
