+++ cocos/physics-2d/framework/index.ts
@@ -1,4 +1,8 @@
+import { legacyCC } from '../../core/global-exports';
+
 import * as PolygonSeparator from './utils/polygon-separator';
+
+import { selector } from './physics-selector';
 
 export * from './physics-types';
 
@@ -28,4 +32,8 @@ export * from './components/joints/hinge-joint-2d';
 
 export const Physics2DUtils = {
     PolygonSeparator,
+};
+
+legacyCC.physics2d = {
+    selector,
 };
