+++ cocos/physics-2d/framework/physics-system.ts
@@ -1,8 +1,8 @@
 import { EDITOR } from 'internal:constants';
 import { System, Vec2, IVec2Like, Rect, Eventify, Enum } from '../../core';
 import { IPhysicsWorld } from '../spec/i-physics-world';
-import { createPhysicsWorld } from './instance';
-import { physicsEngineId } from './physics-selector';
+import { createPhysicsWorld, selector, IPhysicsSelector } from './physics-selector';
+
 import { DelayEvent } from './physics-internal-types';
 import { ICollisionMatrix } from '../../physics/framework/physics-config';
 import { CollisionMatrix } from '../../physics/framework/collision-matrix';
@@ -131,7 +131,9 @@ export class PhysicsSystem2D extends Eventify(System) {
      * @zh
      * 获取物理世界的封装对象，通过它你可以访问到实际的底层对象。
      */
-    readonly physicsWorld: IPhysicsWorld;
+    public get physicsWorld () {
+        return selector.physicsWorld!;
+    }
 
     /**
      * @en
@@ -142,15 +144,15 @@ export class PhysicsSystem2D extends Eventify(System) {
     static readonly ID = 'PHYSICS_2D';
 
     static get PHYSICS_NONE () {
-        return !physicsEngineId;
+        return !selector.id;
     }
 
     static get PHYSICS_BUILTIN () {
-        return physicsEngineId === 'builtin';
+        return selector.id === 'builtin';
     }
 
     static get PHYSICS_BOX2D () {
-        return physicsEngineId === 'box2d';
+        return selector.id === 'box2d';
     }
 
     /**
@@ -229,7 +231,9 @@ export class PhysicsSystem2D extends Eventify(System) {
             }
         }
 
-        this.physicsWorld = createPhysicsWorld();
+        const mutableSelector = selector;
+        mutableSelector.physicsWorld = createPhysicsWorld();
+
         this.gravity = this._gravity;
         this.allowSleep = this._allowSleep;
     }
@@ -347,12 +351,8 @@ export class PhysicsSystem2D extends Eventify(System) {
     }
 }
 
-director.once(Director.EVENT_INIT, () => {
-    initPhysicsSystem();
-});
-
 function initPhysicsSystem () {
-    if (!PhysicsSystem2D.PHYSICS_NONE && (!EDITOR || legacyCC.GAME_VIEW)) {
-        director.registerSystem(PhysicsSystem2D.ID, PhysicsSystem2D.instance, System.Priority.LOW);
-    }
+    director.registerSystem(PhysicsSystem2D.ID, PhysicsSystem2D.instance, System.Priority.LOW);
 }
+
+director.once(Director.EVENT_INIT, () => { initPhysicsSystem(); });
